move voluntaryexits mock
